fix: avoid import conflicts in import strategy, and adapt tests
