+++ .pre-commit-config.yaml
@@ -4,8 +4,10 @@ repos:
     rev: v0.0.280
     hooks:
       - id: ruff
+        exclude: ^tests/fixtures/
   - repo: https://github.com/psf/black
     rev: 23.7.0
     hooks:
       - id: black
         language_version: python3
+        exclude: ^tests/fixtures/
